Change : membuat file untuk masing2 arguments
