+++ lib/models/arguments/module_detail_task_argument.dart
@@ -1,0 +1,7 @@
+class ModuleDetailTaskArguments {
+  final String courseName;
+  final String sectionName;
+  final int numbering;
+
+  ModuleDetailTaskArguments(this.courseName, this.sectionName, this.numbering);
+}
