+++ lib/models/arguments/module_detail_video_argument.dart
@@ -1,0 +1,6 @@
+class ModuleDetailVideoArguments {
+  final String courseName;
+  final String sectionName;
+
+  ModuleDetailVideoArguments(this.courseName, this.sectionName);
+}
