+++ lib/models/arguments/module_list_quiz_argument.dart
@@ -1,0 +1,5 @@
+class ModuleListQuizArguments {
+  final String courseName;
+
+  ModuleListQuizArguments(this.courseName);
+}
