+++ lib/models/arguments/module_screen_argument.dart
@@ -1,0 +1,7 @@
+class ModuleArguments {
+  final String title;
+  final int currentSection;
+  final int totalSection;
+
+  ModuleArguments(this.title, this.currentSection, this.totalSection);
+}
