+++ lib/views/screen/course/module/detail/module_detail_video.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
-import 'package:staredu/models/arguments/module_arguments.dart';
+import 'package:staredu/models/arguments/module_detail_video_argument.dart';
 import 'package:staredu/utils/color/color.dart';
 import 'package:youtube_player_flutter/youtube_player_flutter.dart';
