+++ lib/views/screen/course/module/module_detail_task.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
-import 'package:staredu/models/arguments/module_arguments.dart';
+import 'package:staredu/models/arguments/module_detail_task_argument.dart';
 
 import '../../../../utils/color/color.dart';
